+++ nekoyume/api.py
@@ -5,7 +5,8 @@ from requests import get
 from requests.exceptions import ConnectionError
 from sqlalchemy.exc import IntegrityError
 
-from .models import Block, Move, get_my_public_url
+from .block import Block
+from .models import Move, get_my_public_url
 from .node import Node
 from .orm import db
 from .tasks import block_broadcast, move_broadcast
