+++ nekoyume/cli.py
@@ -8,9 +8,11 @@ from ptpython.repl import embed
 from raven import Client
 
 from .app import app
-from .models import Block, Move, User, get_my_public_url
+from .block import Block
+from .models import Move, get_my_public_url
 from .node import Node
 from .orm import db
+from .user import User
 
 
 DEFAULT_SEED_NODE_URL = os.environ.get(
