+++ tests/game_test.py
@@ -4,8 +4,11 @@ from coincurve import PrivateKey
 from flask.testing import FlaskClient
 from sqlalchemy.orm.session import Session
 
+from nekoyume.block import Block
 from nekoyume.game import get_unconfirmed_move
-from nekoyume.models import Block, Move, User, get_address
+from nekoyume.models import Move
+from nekoyume.user import User
+from nekoyume.util import get_address
 
 
 def test_login(fx_test_client):
